A1.6: derive local semantic execution reachability
